+++ Assets/Scripts/GameCore/Character/Movement/States/MovementHitState.cs
@@ -34,7 +34,7 @@ namespace GameCore.Character.Movement.States
 
             var dataframe = new PlayerEffectStateDataframe
             {
-                type = EffectType.Knockdown,
+                type = (byte)EffectType.Knockdown,
                 active = true,
             };
             movement.GameClient.Send(ref dataframe);
@@ -55,7 +55,7 @@ namespace GameCore.Character.Movement.States
 
             var dataframe = new PlayerEffectStateDataframe
             {
-                type = EffectType.Knockdown,
+                type = (byte)EffectType.Knockdown,
                 active = false,
             };
             movement.GameClient.Send(ref dataframe);
